+++ apps/veil/src/shared/api/server/tournament/social-image/index.tsx
@@ -1,8 +1,11 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { createCanvas } from 'canvas';
-import { drawTournamentEarningsCanvas } from '@/pages/tournament/ui/shared/tournament-earnings-canvas';
+import {
+  renderTournamentEarningsCanvas,
+  TournamentParams,
+  queryParamMap,
+} from '@/features/tournament-earnings-canvas';
 import { registerFonts } from '@/shared/ui/canvas-toolkit';
-import { TournamentParams, queryParamMap } from '@/pages/tournament/ui/join/page';
 
 export async function GET(req: NextRequest) {
   const { searchParams } = new URL(req.url);
@@ -16,7 +19,7 @@ export async function GET(req: NextRequest) {
 
   registerFonts();
   const canvas = createCanvas(600, 315);
-  await drawTournamentEarningsCanvas(canvas as unknown as HTMLCanvasElement, params, true);
+  await renderTournamentEarningsCanvas(canvas as unknown as HTMLCanvasElement, params, true);
 
   return new NextResponse(canvas.toBuffer('image/png'), {
     status: 200,
